requirements.txt
uvicorn==0.24.0.post1
fastapi==0.104.1
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
mysql-connector-python==8.2.0
pydantic==2.5.2
APScheduler~=3.10.4